modules extraction iteration 2
