+++ src/routes/+layout.svelte
@@ -2,37 +2,15 @@
     import "../app.css";
     import Header from "$lib/components/Header.svelte";
     import { onDestroy } from "svelte";
-    import { applyTheme, getSavedTheme, nextTheme, saveTheme, setupSystemThemeListener, type Theme } from "$lib/theme";
+    import { theme, toggleTheme, cleanupTheme } from "$lib/stores/themeStore";
 
     let { children } = $props();
 
-    let theme: Theme = $state(getSavedTheme());
-
-    const systemListener = () => {
-        if (theme === "system") applyTheme(theme);
-    };
-
-    $effect(() => {
-        saveTheme(theme);
-    });
-    $effect(() => {
-        applyTheme(theme);
-    });
-
-    let removeListener: (() => void) | null = null;
-    if (typeof window !== "undefined") {
-        removeListener = setupSystemThemeListener(systemListener);
-    }
-
-    function toggleTheme() {
-        theme = nextTheme(theme);
-    }
-
     onDestroy(() => {
-        removeListener?.();
+        cleanupTheme();
     });
 </script>
 
-<Header {theme} {toggleTheme} />
+<Header theme={$theme} {toggleTheme} />
 
 {@render children()}
